Got a starting point working.
Keyboard handler is located at /dev/input/event# where # is number that might be different on other computers.
The program execute a cat command and pipes in the results.
The output is process using regex to find the correct event handler.
Finally the program opens and closes the handler in question.