program is now successfully reading from keyboard.
main.cpp: Import the input.h file to get the input_event struct.
Reads the from the handler and stores the data in the input_event strcut.
Checks for a EV_KEY and captures the keypress.
Prints the value(press/release) and code (key value) to screen.

README.md: Now contains my thoughts,goals, research and references for the program.

a.out: Removed binaries as the repo should not contain any binaries.