KeyCodes have been mapped.
All keyCodes/scancodes have been mapped to the corresponding keys on keyboard.
Used a vector of strings to store the keycodes. This allows for easier size checking and debugging.
Infinite loop that samples the events going to the API. Exit when ESC is detected.
Tested the program by checking to see if the program correctly detected all keys.