Refactored code and added verbose mode.
Created a header file logger.h with namespace logger. This namespace contains keycode array and three functions.
getDevices() returns string containing a list of devices attached to computer. getHandler() returns a string containing the path to the keyboard handler. connectHandler() connects to the handler and samples the events.

Added verbose mode to the program for better debugging. Verbose mode is turned off by default and can be turned on through the command line using the -v flag.

Added a makefile for easier compiling. In the command line type make to compile the complete project. Type Make clean to rm the binary and remove any .gch files. Binary created is now called keylogger instead of a.out.