Refactored Code.
Moved the code out of the .h file and into a new logger.cpp file. Tested to make sure that it works and nothing is broken. Added header guards to the .h file.

Changed Makefile so that it compiles logger.cpp and logger.h files into an object first before completely compiling the program.